move route handler into client
